ENV/requirements.txt
# === CORE APP ===
streamlit==1.51.0
pandas==2.3.3
numpy==1.26.4
streamlit-modal==0.1.2
email-validator==2.2.0
streamlit-calendar==1.4.0

# === SUPABASE (VERSION ESTABLE + COMPLETA) ===
supabase==2.4.0   # SDK unificado
httpx==0.25.2
gotrue==2.8.1

# === BASE DE DATOS / SQL ===
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# === BACKEND ===
fastapi==0.110.0
uvicorn==0.29.0

# === OPENAI / IA ===
openai==1.51.0

# === VISUALIZACION ===
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0
graphviz==0.20.3
reportlab==4.2.2

# === UTILIDADES ===
python-dotenv==1.0.1
requests==2.32.3
